server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Rapid broker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt==4.3.0",
    "fastapi>=0.119.0",
    "fastapi-localization>=0.0a2.dev3",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.8",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.38.0",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff>=0.14.0",
    "pre-commit<4.1.0,>=4.0.0",
]

[tool.ruff]
exclude = [
    "app/alembic"
]

[tool.pytest.ini_options]
asyncio_mode="auto"
